Alter how captcha images are initialized
This changes the way captcha images get initialized. Previously, the
image was created with the ``PIL.Image.new``[0] method, and then a
background was initialized using the ``PIL.Image.paste``[1] method. The
behaviour of ``paste`` changed in 4.0.0 in a backwards incompatible way.
However, we don't need to use ``paste`` at all, we can pass the
background color directly to ``PIL.Image.new`` and drop the call to
``paste``.

This API has been available since at least Pillow-2.0.0 which ships with
EL7, so this should be compatible with EL7 through Fedora Rawhide
(currently shipping with Pillow-4.0.0).

[0] http://pillow.readthedocs.io/en/3.4.x/reference/Image.html#PIL.Image.new
[1] http://pillow.readthedocs.io/en/3.4.x/reference/Image.html#PIL.Image.Image.paste

fixes #1262

Signed-off-by: Jeremy Cline <jeremy@jcline.org>